Revival checkpoint: Category Editor bugs fixed - used to have a bug where changing Category nam results in new category being created (because category name was used as primary key, and upsert technique is used)

+++ app/src/main/java/com/kaeonx/moneymanager/fragments/transactions/TransactionsRVAdapter.kt
@@ -120,12 +120,12 @@ class TransactionOnClickListener(val clickListener: (transaction: Transaction) -
 
 sealed class RVItem {
     data class RVItemDayTransactions(val dayTransactions: DayTransactions) : RVItem() {
-        override val rvItemId: Long = dayTransactions.dayOfMonth
+        override val rvItemId: Int = dayTransactions.dayOfMonth
     }
 
     object RVItemHeader : RVItem() {
-        override val rvItemId: Long = Long.MIN_VALUE
+        override val rvItemId: Int = Int.MIN_VALUE
     }
 
-    abstract val rvItemId: Long
+    abstract val rvItemId: Int
 }
