+++ app/src/main/java/com/kaeonx/moneymanager/userrepository/domain/DayTransactions.kt
@@ -3,7 +3,7 @@ package com.kaeonx.moneymanager.userrepository.domain
 import java.util.*
 
 data class DayTransactions(
-    val dayOfMonth: Long,  // serves as ID
+    val dayOfMonth: Int,  // serves as ID
     val ymdCalendar: Calendar,
     var dayIncome: String?,  // This value is always in HomeCurrency
     var dayExpenses: String?,  // This value is always in HomeCurrency
